+++ merchantops-api/src/main/java/com/renda/merchantops/api/MerchantOpsApplication.java
@@ -1,5 +1,6 @@
 package com.renda.merchantops.api;
 
+import com.renda.merchantops.api.config.ImportProcessingProperties;
 import com.renda.merchantops.api.config.JwtProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -8,7 +9,7 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication(scanBasePackages = "com.renda.merchantops")
-@EnableConfigurationProperties(JwtProperties.class)
+@EnableConfigurationProperties({JwtProperties.class, ImportProcessingProperties.class})
 @EntityScan(basePackages = "com.renda.merchantops.infra.persistence.entity")
 @EnableJpaRepositories(basePackages = "com.renda.merchantops.infra.repository")
 public class MerchantOpsApplication {